backend/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库
sqlalchemy==2.0.23
alembic==1.12.1
pymysql==1.1.0
cryptography==41.0.7

# Redis
redis==5.0.1

# 异步任务
celery==5.3.4

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6

# HTTP客户端
httpx==0.25.2
requests==2.31.0

# 配置管理
pydantic==2.5.0
pydantic-settings==2.1.0

# 日志
loguru==0.7.2

# 工具库
python-dotenv==1.0.0
click==8.1.7

# AI/ML相关
transformers==4.36.0
torch>=2.2.0
# numpy==1.24.4

# 文件处理
pillow==10.1.0
pdfplumber==0.9.0
python-docx==1.1.0

# 监控
psutil==5.9.6
gpustat==1.1.1

# SSH连接
paramiko==3.5.1

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0